refactor: 스토어명 변경
